+++ src/components/20-molecules/file-upload/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 5.1.0
 
-- Added an `allowedFileType` attribute.
+- Added an `allowedFileTypes` attribute.
 - Now, every file type can be uploaded.
 
 ## 5.0.1